tests/selenium/requirements.txt
selenium==4.15.2
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
Faker==20.1.0
allure-pytest==2.13.2
requests==2.31.0
Flask==3.0.0
Flask-CORS==4.0.0
